Allow search parameters to use a long and not int

+++ api/src/main/java/uk/ac/ox/oucs/search2/SearchService.java
@@ -13,9 +13,9 @@ public interface SearchService {
 
     SearchResultList search(String searchQuery, Collection<String> contexts);
 
-    SearchResultList search(String searchQuery, int start, int length);
+    SearchResultList search(String searchQuery, long start, long length);
 
-    SearchResultList search(String searchQuery, Collection<String> contexts, int start, int length);
+    SearchResultList search(String searchQuery, Collection<String> contexts, long start, long length);
 
     String getSuggestion(String searchString);
