+++ backward-compatibility/src/main/java/uk/ac/ox/oucs/search2/backwardcompatibility/BackSearchService.java
@@ -18,9 +18,9 @@ public class BackSearchService extends AbstractSearchService {
     private SearchService searchService;
 
     @Override
-    protected SearchResultList search(String searchQuery, Collection<String> contexts, int start, int length, Iterable<SearchFilter> filterChain) {
+    protected SearchResultList search(String searchQuery, Collection<String> contexts, long start, long length, Iterable<SearchFilter> filterChain) {
         try {
-            SearchList searchList = searchService.search(searchQuery, new ArrayList<String>(contexts), start, start + length, null, null);
+            SearchList searchList = searchService.search(searchQuery, new ArrayList<String>(contexts), (int) start, (int) (start + length), null, null);
             return new BackSearchResultList(searchList, filterChain);
         } catch (InvalidSearchQueryException e) {
             //TODO: Log that
